fix(benchmark): make httpbench compile on Windows
native-ci runs `go test ./...` on every module, including this one, on
Windows — where syscall.SysProcAttr.Setpgid / syscall.Getpgid / syscall.Kill
do not exist, so the package failed to build. Move the process-group setup
and teardown behind build tags: the real unix implementation (proc_unix.go)
and a compile-only Kill fallback (proc_other.go) for non-unix platforms.
httpbench only ever runs on unix (Linux CI / dev hosts); the fallback just
keeps the package buildable so the cross-platform test pass is green.
